Create multiple_of_3and5
If we list all the natural numbers below 10 that are multiples of 3 or 5, we get  3, 5, 6and 9. The sum of these multiples is 23.

Find the sum of all the multiples of  3 or 5.

Sample input:
1.  2
2.  10
3.  100

Sample output:
1.  23
2.  2318


For N=10, if we list all the natural numbers below 10 that are multiples of  3 or 5, we get 3, 5, 6 and 9. The sum of these multiples is 23.

Similarly for N=100, we get 2318.